move Metric protos from model/fn-execution to model/pipeline
